+++ tester_cpp/source/tested/ft_strlen.cpp
@@ -52,7 +52,7 @@ TEST(ft_strlen_random_string) {
     TestResult res{};
     res.n_tested = TestSize::M;
 
-    for (int i = 0; i < res.n_tested; ++i) {
+    for (size_t i = 0; i < res.n_tested; ++i) {
         std::string s = generateRandomString(rng, TestSize::S);
         size_t expected = strlen(s.c_str());
         size_t got = ft_strlen(s.c_str());
